Replace np.bool, np.int, np.float with np.bool_, np.int_, np.float_ (#5103)
* Replace np.bool with np.bool_ everywhere

* Replace np.int to np.int_ and np.float with np.float_

* Fix PEP8 issues

* Change np.type_ to type

* Fix incorrect find-and-replace of cnp.float_t

* Add all dtype aliases to convert types

* use .astype instead of cast

+++ skimage/exposure/tests/test_exposure.py
@@ -245,7 +245,7 @@ def test_rescale_out_range():
     """
     image = np.array([-10, 0, 10], dtype=np.int8)
     out = exposure.rescale_intensity(image, out_range=(0, 127))
-    assert out.dtype == np.float_
+    assert out.dtype == float
     assert_array_almost_equal(out, [0, 63.5, 127])
 
 
@@ -326,7 +326,7 @@ def test_rescale_nan_warning(in_range, out_range):
         ('uint10', np.uint16),
         ('uint12', np.uint16),
         ('uint16', np.uint16),
-        ('float', np.float_),
+        ('float', float),
     ]
 )
 def test_rescale_output_dtype(out_range, out_dtype):
@@ -346,7 +346,7 @@ def test_rescale_float_output():
     image = np.array([-128, 0, 127], dtype=np.int8)
     output_image = exposure.rescale_intensity(image, out_range=(0, 255))
     testing.assert_array_equal(output_image, [0, 128, 255])
-    assert output_image.dtype == np.float_
+    assert output_image.dtype == float
 
 
 def test_rescale_raises_on_incorrect_out_range():
